Add files via upload
had "watch dog" restarts to often. Figured it was from a couple different reasons. The webpages displaying time in seconds was keeping the processor busy. I removed displaying seconds and now call for updates only every minute.
The checkAP function was trying to reconnect the network, or check for connection to often. I removed both of those. 
Changed settings selection for time to live to a minimum of 3 minutes. Found that less than that AP would time out before settings could be accessed and changed.